requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
groq==0.4.1
Werkzeug==2.3.7
